TASK: Introduce limited retry with exponential backoff
The default beavior still uses no delay and tries an infinited number
of times to not break existing settings.

A new configuration option allow for linear and exponential backoff
strategy, which means a delay is either added with a fixed number of
seconds or with exponential growth.

A second new configuration option allows to set the number of retries.