feat: Force normal execution
Manually calling CopyToScheduler::forceNormalExecution(callable) creates
an process scope where even ScheduledJobInterface::execute() would always
be called immediately without going to the scheduler.

This is meant to be used in test scenarios, to avoid Scheduler::next()
calls wherever the job being scheduled or not isn't a thing we want
to test for.