fix: Prevent parallel execution of re-scheduled jobs
This adds a "running" flag to jobs, in addition to the already existing
"claimed" tag. Jobs now get "claimed" as well as marked as "running" when
working on them starts. Re-scheduling a claimed job empties the "claimed"
tag but keeps the "running" flag as is. This results in  a "one more,
please" situation but prevents the additional job from being claimed again
as long as the previous one is still being worked on.