fix: Consider "running" flag when applying backoff
When a job is re-scheduled, there's two candidates for due
dates, the one already known in the jobs table and the
one which is on its way in.

Without the patch, always the lower value of both is used.
This breaks the backoff mechanism, because in case of
retries the backoff value will always be higher and thus
never used.

When a job is re-scheduled which are currently not running,
the lower due date between the one already known and the
one getting re-added musted be used. That's the behavior
making scheduled jobs "debounce".

When a job is re-scheduled which is currently running, the
due date of the new job must be used, even if it's not the
lower value. This fixes the backoff behavior.

Related: #14