fix: Don't unserialize jobs unless they are executed
The previous implementation relied on Doctrine unserializing
properties with "type=object", which was used for ScheduledJob::$job.

This change passes the serialized data to the React/Process calling
the ./flow job:execute request to prevent having any unserialized
job object in the scheduler loop, because unserialization might load
DBAL connections and thus interact poorly with timeouts.