fix: Reset jobs even if they are unclaimed
The "claimed" marker only indicates if there is currently a worker
handling a job if that job hasn’t been rescheduled. In case of
rescheduling, the "claimed" marker gets set to "", but the "running"
state stays 1.

This means resetting the "running" state of stale jobs must not
omit unclaimed jobs.